+++ example/android/app/src/main/kotlin/com/anntech/example/ann_flavor_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.anntech.example.ann_flavor_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()